unscroll/requirements.txt
appdirs==1.4.3
args==0.1.0
autopep8==1.3.1
base==1.0.4
beautifulsoup4==4.5.3
bs4==0.0.1
click==6.7
clint==0.5.1
convertdate==2.1.1
coreapi==2.3.0
coreapi-cli==1.0.6
coreschema==0.0.4
dateparser==0.6.0
defusedxml==0.5.0
Django==1.11
django-allauth==0.31.0
django-crispy-forms==1.6.1
django-filter==1.0.2
django-filters==0.2.1
django-rest-auth==0.9.1
django-rest-swagger==2.1.2
django-silk==1.0.0
djangorestframework==3.6.2
djangorestframework-bulk==0.2.1
docopt==0.6.2
ephem==3.7.6.0
importmagic==0.1.7
internetarchive==1.5.0
itypes==1.1.0
jdatetime==1.8.2
jedi==0.10.2
Jinja2==2.9.6
jsonpatch==1.15
jsonpointer==1.10
MarkupSafe==1.0
oauthlib==2.0.2
olefile==0.44
openapi-codec==1.3.1
packaging==16.8
pbr==2.1.0
pep8==1.7.0
Pillow==4.1.0
pip-review==0.5.3
pycodestyle==2.3.1
Pygments==2.2.0
pyparsing==2.2.0
python-baseconv==1.2.0
python-dateutil==2.6.0
python3-openid==3.1.0
pytz==2017.2
regex==2017.4.5
requests==2.13.0
requests-oauthlib==0.8.0
rope-py3k==0.9.4.post1
ruamel.yaml==0.14.6
schema==0.6.5
simplejson==3.10.0
six==1.10.0
sqlparse==0.2.3
stevedore==1.21.0
tzlocal==1.3.2
umalqurra==0.2
uritemplate==3.0.0
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
yapf==0.16.1
